+++ src/public/Invoke-SubDirectoryGitCommand.ps1
@@ -10,6 +10,9 @@ directories.
 .PARAMETER Command
 The Git command to execute, excluding the Git executable.
 
+.PARAMETER CommandText
+The text of the Git command to execute, excluding the Git executable.
+
 .PARAMETER Path
 The parent directory to find all child Git repositories from.
 
@@ -34,7 +37,7 @@ Gets the value that determines whether commit are gpg-signed for a directory,
 hiding the head and ahead/behind values for each repository.
 #>
 function Invoke-SubDirectoryGitCommand {
-    [CmdletBinding()]
+    [CmdletBinding(DefaultParameterSetName = 'String')]
     [Diagnostics.CodeAnalysis.SuppressMessageAttribute(
         'PSReviewUnusedParameter',
         'Command',
@@ -48,25 +51,45 @@ function Invoke-SubDirectoryGitCommand {
         'NoAheadBehind',
         Justification = 'False positive since rule does not scan child scopes')]
     param(
-        [Parameter(Mandatory, ValueFromPipeline)]
+        [Parameter(Mandatory, ValueFromPipeline, ParameterSetName = 'String', Position = 0)]
+        [string]$CommandText,
+
+        [Parameter(Mandatory, ValueFromPipeline, ParameterSetName = 'Script', Position = 0)]
         [scriptblock]$Command,
-        [Parameter()]
+
+        [Parameter(ParameterSetName = 'Script', Position = 1)]
+        [Parameter(ParameterSetName = 'String', Position = 1)]
         [string]$Path = '.',
+
+        [Parameter(ParameterSetName = 'Script')]
+        [Parameter(ParameterSetName = 'String')]
         [switch]$NoHead,
+
+        [Parameter(ParameterSetName = 'Script')]
+        [Parameter(ParameterSetName = 'String')]
         [switch]$NoAheadBehind
     )
     process {
+        if ($Command) {
+            $CommandText = $Command.ToString()
+        }
+
         Get-ChildItem -Path $Path -Directory |
             ForEach-Object {
+                $gitDirectory = Join-Path -Path $_ -ChildPath '.git'
+                if (-not (Test-Path -Path $gitDirectory)) {
+                    return
+                }
+
                 Write-GitRepositoryDetail -Path $_.FullName -NoHead:$NoHead -NoAheadBehind:$NoAheadBehind
 
-                $trimmedCommand = $Command.ToString().Trim()
-                if ($trimmedCommand -like 'noop*') {
+                $trimmedCommandText = $CommandText.Trim()
+                if ($trimmedCommandText -like 'noop*') {
                     # Do not show the current branch via the provided command since it is displayed along with the directory name.
                     return
                 }
 
-                $executedCommand = "git -C '$($_.FullName)' $trimmedCommand"
+                $executedCommand = "git -C '$($_.FullName)' $trimmedCommandText"
                 Invoke-Command -ScriptBlock ([scriptblock]::Create($executedCommand))
             }
     }