+++ CHANGELOG.md
@@ -7,7 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.1.0]
+### Added
+
+- Allow text-based sub-directory Git commands.
+- Support JSON output for .NET target framework version check.
+
+### Changed
+
+- Skip Git command execution in non-Git directories.
+- Remove dot path segment from relative output in .NET target framework check.
+- Update .NET support policy for target framework check.
+
+## [0.1.0] - 2024-06-20
 
 ### Added
 
@@ -20,5 +31,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add Get-DotnetTargetFramework cmdlet (#4).
 - Add Get-DotnetOutdatedPackage cmdlet (#5).
 
-[unreleased]: https://github.com/lancra/pwsh/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/lancra/pwsh/compare/v0.1.0...HEAD
 [0.1.0]: https://github.com/lancra/pwsh/releases/tag/v0.1.0